chore: correct example file
